Refactor into modules
The giant `lib.rs` had a few inline submodules already, which are now
broken out into separate module files. I've sliced a few other groups of
functionality into new modules here as well.

There should be no functional changes here.
